refactor: split three oversized files under the shape caps (#2004)
Pure reorganization — no behavior change.

- db/src/cross_format.rs (1153 lines) becomes a thin module file over
  link / mapping / declare / resume / alignment / order sub-modules;
  resume_candidate, alignment_view and declare_against_link are broken
  into named helpers each under the 80-line cap.
- frontend/src/components/alignment_modal.rs (1001 lines) keeps the modal
  shell and moves its wording, lanes, and multi-file choice into
  alignment_modal/{copy,lanes,choice}.rs.
- frontend/src/pages/listen/bootstrap/mod.rs (865 lines) moves the JS
  callbacks into callbacks.rs and the manifest branch into manifest.rs;
  js.rs splits its one oversized listener string into four named
  per-concern builders that concatenate byte-for-byte to the original.
