+++ frontend/src/pages/listen/bootstrap/mod.rs
@@ -2,19 +2,23 @@
 //! `window`, wires Rust callbacks for time / duration / play / pause, then
 //! fetches the manifest and branches on `mode` (direct vs hls) to seed
 //! playback. Extracted from `BookListenPage` so the parent stays under the
-//! 150-line component cap; everything here is web-feature gated.
+//! 150-line component cap; everything here is web-feature gated. The
+//! callback closures live in [`callbacks`], the manifest branch in
+//! [`manifest`], and the JS strings in [`js`].
 
 #![cfg(feature = "web")]
 
 use dioxus::prelude::*;
-use wasm_bindgen::prelude::*;
 
-use super::helpers::{post_audio_progress, resolve_boot_file, resolve_boot_position};
 use crate::data;
 
+mod callbacks;
 mod js;
+mod manifest;
 
-use js::{control_surface_js, eval_hls_init, inject_hls_script};
+use callbacks::{register_js_callbacks, register_readiness_callbacks};
+use js::{control_surface_js, inject_hls_script};
+use manifest::run_manifest_init;
 
 /// Owns the `Closure`s wired into `window.__omnibusOn*` callbacks. Held
 /// across re-renders by `use_hook` so the closures outlive the effect
@@ -294,240 +298,6 @@ fn spawn_manifest_init(
     });
 }
 
-/// Wire the six Rust-side closures that JS calls back into:
-/// `__omnibusOnAudioTime`, `__omnibusOnAudioDuration`, `__omnibusOnAudioPlay`,
-/// `__omnibusOnAudioPause`, `__omnibusOnAudioEnded`, `__omnibusOnInitTimeout`.
-/// Registered before the JS bootstrap so a fast `loadedmetadata` always finds
-/// them.
-///
-/// `loaded_file_id` is the shared signal `run_manifest_init` fills once the
-/// manifest names the file it resolved; every position POST reads it live,
-/// so writes name the file playback is actually in even though these
-/// closures are registered before the manifest lands (#1888). A file switch
-/// always re-boots (see [`needs_reload`]), which resets the signal before
-/// the next manifest fills it again.
-#[expect(clippy::too_many_arguments, reason = "one closure per JS callback")]
-fn register_js_callbacks(
-    cb_holder: &JsCallbackHolder,
-    uuid_cb: String,
-    loaded_file_id: Signal<Option<i64>>,
-    mut duration: Signal<f64>,
-    mut elapsed: Signal<f64>,
-    mut playing: Signal<bool>,
-    mut playback_failed: Signal<bool>,
-    file_sig: Signal<Option<i64>>,
-    uuid_sig: Signal<Option<String>>,
-) {
-    let Some(window) = web_sys::window() else {
-        return;
-    };
-    let uuid_for_save = uuid_cb.clone();
-    // Shared between the heartbeat and the seek path so a seek resets the
-    // heartbeat's 5s window rather than double-writing moments later.
-    let last_saved = std::rc::Rc::new(std::cell::Cell::new(0.0_f64));
-    let on_time = {
-        let last_saved = last_saved.clone();
-        let uuid_for_save = uuid_for_save.clone();
-        // `user_acted` is the shim's `_userActed`: false until a real play
-        // or user seek this boot. The boot restore sets `currentTime`,
-        // which fires a `timeupdate` — persisting that write re-clocked an
-        // unmoved row on every visit and shadowed the counterpart at the
-        // cross-format clock gate (#1972). Render the time; persist only
-        // positions a human produced.
-        Closure::<dyn FnMut(f64, bool)>::new(move |secs: f64, user_acted: bool| {
-            elapsed.set(secs);
-            if !user_acted {
-                return;
-            }
-            if (secs - last_saved.get()).abs() < 5.0 {
-                return;
-            }
-            last_saved.set(secs);
-            crate::audiobook_progress::save(&uuid_for_save, secs);
-            post_audio_progress(uuid_for_save.clone(), *loaded_file_id.peek(), secs);
-        })
-    };
-    // A seek is discrete user intent — persist it immediately, bypassing
-    // the heartbeat throttle. A paused element fires no further events, so
-    // waiting on the next timeupdate loses the seek if the page exits
-    // first (#1972).
-    let on_seeked = {
-        let last_saved = last_saved.clone();
-        let uuid_for_seek = uuid_for_save.clone();
-        Closure::<dyn FnMut(f64)>::new(move |secs: f64| {
-            elapsed.set(secs);
-            last_saved.set(secs);
-            crate::audiobook_progress::save(&uuid_for_seek, secs);
-            post_audio_progress(uuid_for_seek.clone(), *loaded_file_id.peek(), secs);
-        })
-    };
-    let on_duration = Closure::<dyn FnMut(f64)>::new(move |d: f64| {
-        duration.set(d);
-    });
-    // Starting an `Unread` book marks it `Reading`, the audio counterpart of
-    // the readers' mark-on-open. Fired from the first play of this boot only
-    // — `play` also fires on every resume from pause, and a file switch
-    // re-registers these callbacks, so the flag scopes it to one write per
-    // book opened.
-    let uuid_for_start = uuid_cb.clone();
-    let mut marked_reading = false;
-    let on_play = Closure::<dyn FnMut(f64)>::new(move |_: f64| {
-        playing.set(true);
-        if marked_reading {
-            return;
-        }
-        marked_reading = true;
-        let uuid = uuid_for_start.clone();
-        wasm_bindgen_futures::spawn_local(async move {
-            crate::read_status_auto::apply_auto_read_status("", &uuid, false).await;
-        });
-    });
-    // This FILE played through (the `ended` handler in `js.rs` only calls
-    // once there is no next part). On a multi-file sequence the book is
-    // not done — advance the player to the next file's start (paused;
-    // the boot leaves it unseeded until a real play) instead of marking
-    // five-hours-in as finished. Only the last file finishes the book.
-    let uuid_for_end = uuid_cb.clone();
-    let on_ended = Closure::<dyn FnMut(f64)>::new(move |_: f64| {
-        if let Some(next) = NEXT_SEQUENCE_FILE.with(|c| c.get()) {
-            let uuid = uuid_for_end.clone();
-            let mut file_sig = file_sig;
-            let mut uuid_sig = uuid_sig;
-            file_sig.set(Some(next));
-            // The driver keys on the uuid changing; a None→Some pair in
-            // one render collapses to no-change, so the Some lands from a
-            // spawned task after the None commits.
-            uuid_sig.set(None);
-            wasm_bindgen_futures::spawn_local(async move {
-                uuid_sig.set(Some(uuid));
-            });
-            return;
-        }
-        let uuid = uuid_for_end.clone();
-        wasm_bindgen_futures::spawn_local(async move {
-            crate::read_status_auto::apply_auto_read_status("", &uuid, true).await;
-        });
-    });
-    let uuid_for_pause = uuid_cb;
-    // The playing flag flips on every pause (src swaps and teardown fire
-    // them too), but only a session the user actually drove persists its
-    // position — a programmatic pause on an untouched boot re-states the
-    // seeded spot with a fresh clock (#1972).
-    let on_pause = Closure::<dyn FnMut(f64, bool)>::new(move |secs: f64, user_acted: bool| {
-        playing.set(false);
-        if !user_acted {
-            return;
-        }
-        crate::audiobook_progress::save(&uuid_for_pause, secs);
-        post_audio_progress(uuid_for_pause.clone(), *loaded_file_id.peek(), secs);
-    });
-    // Fired from the init-poll's `n >= 200` branch when
-    // `window.OmnibusAudio` never appears (mount loop gave
-    // up, JS eval failed, vendored asset 404'd, …). Mirrors
-    // the `reader.rs::__omnibusOnStatus("error")` shape so
-    // the UI surfaces a real failure instead of stalling
-    // on a perpetually-not-ready state.
-    let on_init_timeout = Closure::<dyn FnMut(f64)>::new(move |_: f64| {
-        playback_failed.set(true);
-    });
-
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioTime"),
-        on_time.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioSeeked"),
-        on_seeked.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioDuration"),
-        on_duration.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioPlay"),
-        on_play.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioPause"),
-        on_pause.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioEnded"),
-        on_ended.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnInitTimeout"),
-        on_init_timeout.as_ref().unchecked_ref(),
-    );
-    *cb_holder.borrow_mut() = vec![
-        Box::new(on_time),
-        Box::new(on_seeked),
-        Box::new(on_duration),
-        Box::new(on_play),
-        Box::new(on_pause),
-        Box::new(on_ended),
-        Box::new(on_init_timeout),
-    ];
-}
-
-/// Wire the three JS→Rust readiness/health callbacks:
-/// `__omnibusOnAudioBooted` fires once `initDirect`/`initHls` has actually
-/// committed a source to the element — the authoritative "ready" signal,
-/// replacing the old optimistic `hls_ready.set(true)` issued right after
-/// firing the init eval (a real, if usually tiny, race against the JS
-/// actually running). `__omnibusOnAudioBuffering` mirrors the JS
-/// watchdog's `readyState < HAVE_FUTURE_DATA` check while playing.
-/// `__omnibusOnAudioStalled` reuses `playback_failed` — a stall with no
-/// forward progress is exactly as unrecoverable-without-a-reload as an
-/// init timeout. Closures are appended to `cb_holder` alongside
-/// `register_js_callbacks`'s so both sets outlive the installing effect.
-fn register_readiness_callbacks(
-    cb_holder: &JsCallbackHolder,
-    mut hls_ready: Signal<bool>,
-    mut buffering: Signal<bool>,
-    mut playback_failed: Signal<bool>,
-) {
-    let Some(window) = web_sys::window() else {
-        return;
-    };
-    let on_booted = Closure::<dyn FnMut(f64)>::new(move |_: f64| {
-        hls_ready.set(true);
-    });
-    let on_buffering = Closure::<dyn FnMut(f64)>::new(move |flag: f64| {
-        buffering.set(flag > 0.5);
-    });
-    let on_stalled = Closure::<dyn FnMut(f64)>::new(move |_: f64| {
-        playback_failed.set(true);
-    });
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioBooted"),
-        on_booted.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioBuffering"),
-        on_buffering.as_ref().unchecked_ref(),
-    );
-    let _ = js_sys::Reflect::set(
-        &window,
-        &JsValue::from_str("__omnibusOnAudioStalled"),
-        on_stalled.as_ref().unchecked_ref(),
-    );
-    cb_holder.borrow_mut().extend([
-        Box::new(on_booted) as Box<dyn std::any::Any>,
-        Box::new(on_buffering),
-        Box::new(on_stalled),
-    ]);
-}
-
 /// Install the `window.OmnibusAudio` control surface immediately so
 /// the transport buttons are wired even before `initDirect` / `initHls`
 /// attaches. The two init paths are responsible for setting their own
@@ -546,282 +316,6 @@ thread_local! {
     /// between sequence-advance and marking the book finished. Reset at
     /// every boot entry so a failed load can't leave a stale advance.
     static NEXT_SEQUENCE_FILE: std::cell::Cell<Option<i64>> = const { std::cell::Cell::new(None) };
-}
-
-/// Fetch `/api/audiobooks/{uuid}/manifest` — targeting the picker's
-/// `file_id` when one was selected, else the progress row's stored
-/// `book_file_id` so multi-file books resume in the file the seconds were
-/// recorded in (#1888) — then either:
-/// * **Direct mode** — call `initDirect` with the parts list (instant
-///   playback for m4b/m4a/mp3/aac). `hls_ready` flips once the JS side
-///   confirms via `__omnibusOnAudioBooted` that a source is actually
-///   committed, not optimistically here.
-/// * **HLS mode** — poll `/status` until `ready` (call `initHls`, same
-///   JS-confirmed `hls_ready` flip) or `failed` (flip `playback_failed`).
-/// * **No manifest** — show the same failure overlay as a terminal HLS
-///   transcode failure.
-async fn run_manifest_init(
-    uuid_for_fetch: String,
-    file_id: Option<i64>,
-    user_id: Option<i64>,
-    initial_position: f64,
-    playback: crate::PlaybackState,
-    current_user: Signal<Option<Option<omnibus_shared::UserSummary>>>,
-) {
-    let mut rate = playback.rate;
-    let mut rate_error = playback.rate_error;
-    let mut playback_failed = playback.playback_failed;
-    let chapters_sig = playback.chapters;
-    let uuid_guard = playback.uuid;
-    // True only while `uuid_for_fetch` is still the active book. Checked before
-    // every shared-signal write so a stale task (user switched books mid-fetch
-    // or mid-`/status`-poll) can't clobber the new book's state.
-    let is_current = || {
-        let active_user_id = current_user
-            .peek()
-            .as_ref()
-            .and_then(|user| user.as_ref())
-            .map(|user| user.id);
-        crate::audiobook_progress::playback_load_matches(
-            uuid_guard.peek().as_deref(),
-            active_user_id,
-            &uuid_for_fetch,
-            user_id,
-        )
-    };
-    // Reconcile resume state with the server upfront: the row carries both
-    // the seconds and the `book_files` row they were recorded in, and the
-    // file decides which manifest to fetch before any position applies.
-    let server_row = data::get_progress("", &uuid_for_fetch, omnibus_shared::ProgressFormat::Audio)
-        .await
-        .ok()
-        .flatten();
-
-    let server_rate = data::get_playback_rate("", &uuid_for_fetch).await;
-    if !is_current() {
-        return;
-    }
-    if let (Some(user_id), Ok(Some(record))) = (user_id, server_rate.as_ref()) {
-        crate::audiobook_progress::save_rate(user_id, &uuid_for_fetch, record.playback_rate);
-    }
-    let local_rate =
-        user_id.and_then(|id| crate::audiobook_progress::load_rate(id, &uuid_for_fetch));
-    let resolution = crate::audiobook_progress::resolve_rate(
-        server_rate
-            .as_ref()
-            .map(|record| record.as_ref().map(|record| record.playback_rate))
-            .map_err(|_| ()),
-        local_rate,
-    );
-    rate.set(resolution.playback_rate);
-    super::helpers::audio_call("setRate", &resolution.playback_rate.to_string());
-    if resolution.seed_server {
-        let update = omnibus_shared::AudiobookPlaybackRateUpdate {
-            playback_rate: resolution.playback_rate,
-        };
-        if let Err(error) = data::set_playback_rate("", &uuid_for_fetch, update).await {
-            if is_current() {
-                rate_error.set(Some(format!("Could not save playback speed: {error}")));
-            }
-        }
-    }
-
-    // Resolve follow at boot: when the link follows the counterpart and the
-    // reading position maps newer, seed playback with the mapped spot
-    // instead of the stored row — a post-boot corrective seek would race
-    // `initDirect`'s one-shot restore seek and lose. Fetched even for an
-    // explicit `?file_id=` — every Continue card carries the row's file id,
-    // and skipping follow there reopened books at the stale spot (#1972);
-    // `resolve_follow_boot` still stands aside when the pick names a
-    // *different* file than the candidate maps to.
-    let follow_resume =
-        super::sync_prompt::fetch_resume_with(&uuid_for_fetch, "audio", false).await;
-    if !is_current() {
-        return;
-    }
-    let follow_boot = super::helpers::resolve_follow_boot(file_id, follow_resume.as_ref());
-
-    let row_file = server_row.as_ref().and_then(|r| r.book_file_id);
-    let boot_file = follow_boot
-        .as_ref()
-        .and_then(|(file, _)| *file)
-        .or_else(|| resolve_boot_file(file_id, row_file));
-    let mut manifest = fetch_manifest(&uuid_for_fetch, boot_file).await;
-    // The row's stored id is a soft reference — a reindex may have replaced
-    // the `book_files` row since the position was saved. Fall back to the
-    // server's default file rather than a failure overlay, mirroring
-    // `db::progress::resume_points`. Never applied to an explicit picker
-    // selection: a dead `?file_id=` should fail loudly, not play another file.
-    if manifest.is_none() && file_id.is_none() && boot_file.is_some() {
-        manifest = fetch_manifest(&uuid_for_fetch, None).await;
-    }
-
-    // A newer book may have been selected while the manifest was in flight.
-    // The Direct + None arms below are synchronous, so this single guard covers
-    // them; the HLS arm re-checks each poll iteration.
-    if !is_current() {
-        return;
-    }
-
-    let Some(manifest) = manifest else {
-        // Manifest unreachable (network failure, 5xx,
-        // 404 between settings save and reindex). Show
-        // the same failure overlay as a terminal HLS
-        // transcode failure — a manual refresh is the
-        // recovery path either way.
-        playback_failed.set(true);
-        return;
-    };
-
-    let (loaded_file, audio_file_count) = manifest.file_identity();
-    // Sequence auto-advance is a DIRECT-mode contract only: the HLS
-    // playlist is book-level and file-blind (`get_audiobook_playlist`
-    // takes no file id), so "advancing" an HLS book reboots the same
-    // stream at 0:00 forever instead of finishing it. HLS keeps the
-    // ended-means-finished behavior.
-    let next_file = match &manifest {
-        omnibus_shared::AudiobookManifest::Direct { .. } => manifest.next_file_id(),
-        omnibus_shared::AudiobookManifest::Hls { .. } => None,
-    };
-    NEXT_SEQUENCE_FILE.with(|c| c.set(next_file));
-    // `0` = a server predating the identity fields; keep posting no file id
-    // rather than inventing one (`resolve_boot_position` degrades the same
-    // way via `audio_file_count == 0`).
-    let mut loaded_file_sig = playback.loaded_file_id;
-    loaded_file_sig.set((loaded_file > 0).then_some(loaded_file));
-    // The mapped position applies only when the manifest actually loaded the
-    // candidate's file (the dead-row fallback above may have swapped it) —
-    // otherwise fall back to the stored-row resolution unchanged.
-    let follow_pos = follow_boot
-        .as_ref()
-        .filter(|(file, _)| match file {
-            Some(fid) => *fid == loaded_file || loaded_file == 0,
-            None => true,
-        })
-        .map(|(_, seconds)| *seconds);
-    // `None` = unseeded: `initDirect` gets a JSON `null`, applies no seek,
-    // and leaves the transport's seed gate closed — an unattributable boot
-    // must stay unpersistable until the user really plays or seeks.
-    let resume_pos = follow_pos.or_else(|| {
-        resolve_boot_position(
-            server_row
-                .as_ref()
-                .map(|r| (r.book_file_id, r.audio_position_seconds)),
-            initial_position,
-            loaded_file,
-            audio_file_count,
-        )
-    });
-    let pos_lit = serde_json::to_string(&resume_pos).unwrap_or_else(|_| "null".into());
-
-    match manifest {
-        omnibus_shared::AudiobookManifest::Direct {
-            parts, chapters, ..
-        } => init_direct_play(parts, chapters, &pos_lit, chapters_sig),
-        omnibus_shared::AudiobookManifest::Hls { playlist_url, .. } => {
-            init_hls(
-                &uuid_for_fetch,
-                &playlist_url,
-                &pos_lit,
-                is_current,
-                playback_failed,
-            )
-            .await;
-        }
-    }
-}
-
-/// GET the audiobook manifest (`?file_id=N` for multi-file books) and
-/// decode it. Returns `None` on network error or non-200.
-async fn fetch_manifest(
-    uuid: &str,
-    file_id: Option<i64>,
-) -> Option<omnibus_shared::AudiobookManifest> {
-    let manifest_url = match file_id {
-        Some(fid) => format!("/api/audiobooks/{uuid}/manifest?file_id={fid}"),
-        None => format!("/api/audiobooks/{uuid}/manifest"),
-    };
-    match gloo_net::http::Request::get(&manifest_url).send().await {
-        Ok(resp) if resp.status() == 200 => resp.json().await.ok(),
-        _ => None,
-    }
-}
-
-/// Direct-mode arm: hand the part list to the JS `initDirect` shim and
-/// publish the chapter map. Synchronous — the caller's `is_current` guard
-/// covered the only `await`. `hls_ready` is no longer flipped here — the
-/// JS shim reports back via `__omnibusOnAudioBooted` once it has actually
-/// committed a source, closing the race where a click could fire
-/// `el.play()` before `el.src` existed.
-fn init_direct_play(
-    parts: Vec<omnibus_shared::ManifestPart>,
-    chapters: Vec<omnibus_shared::ChapterInfo>,
-    pos_lit: &str,
-    mut chapters_sig: Signal<Vec<omnibus_shared::ChapterInfo>>,
-) {
-    // Populate chapter signal from manifest data.
-    chapters_sig.set(chapters);
-
-    // Hand the part list to JS; initDirect picks
-    // the right starting part by cumulative offset.
-    let parts_json = serde_json::to_string(&parts).unwrap_or_else(|_| "[]".into());
-    let init_js = format!(
-        r#"(function(){{ var n=0; (function go(){{ if (window.OmnibusAudio) {{ window.OmnibusAudio.initDirect({parts_json}, {pos_lit}); }} else if (n++ < 200) {{ setTimeout(go, 50); }} else {{ console.error('OmnibusAudio never installed; init timed out'); if (typeof window.__omnibusOnInitTimeout === 'function') {{ window.__omnibusOnInitTimeout(0); }} }} }})(); }})();"#
-    );
-    let _ = dioxus::document::eval(&init_js);
-}
-
-/// HLS arm: poll `/api/audiobooks/{uuid}/status` until `ready` (call
-/// `initHls`, which reports back via `__omnibusOnAudioBooted` the same way
-/// the direct-mode arm does) or `failed` (flip `playback_failed`).
-/// Re-checks `is_current` every iteration so a stale poll can't clobber
-/// the newly-active book's signals.
-async fn init_hls(
-    uuid: &str,
-    playlist_url: &str,
-    pos_lit: &str,
-    is_current: impl Fn() -> bool,
-    mut playback_failed: Signal<bool>,
-) {
-    let playlist_lit = serde_json::to_string(playlist_url).unwrap_or_else(|_| "\"\"".into());
-    loop {
-        // Stop polling the moment the user switches away from this book,
-        // so a stale `/status` loop can't flip the new book's signals.
-        if !is_current() {
-            return;
-        }
-        match fetch_hls_status(uuid).await.as_deref() {
-            // Bug 4 from #338: surface failed transcodes instead of
-            // polling forever.
-            Some("failed") => {
-                playback_failed.set(true);
-                return;
-            }
-            Some("ready") => {
-                eval_hls_init(&playlist_lit, pos_lit);
-                return;
-            }
-            _ => {}
-        }
-        gloo_timers::future::TimeoutFuture::new(1_000).await;
-    }
-}
-
-/// One HLS `/status` poll: fetch + decode the JSON body and return the
-/// `state` field, or `None` on network / decode failure (the caller
-/// just keeps polling).
-async fn fetch_hls_status(uuid: &str) -> Option<String> {
-    let resp = gloo_net::http::Request::get(&format!("/api/audiobooks/{uuid}/status"))
-        .send()
-        .await
-        .ok()?;
-    if resp.status() != 200 {
-        return None;
-    }
-    let json: serde_json::Value = resp.json().await.ok()?;
-    json.get("state")
-        .and_then(|v| v.as_str())
-        .map(str::to_owned)
 }
 
 // Tests for the pure Rust-side logic only. The JS interop seams
